services/auth-service/requirements.txt
fastapi
uvicorn
python-jose
passlib[bcrypt]
azure-cosmos
pydantic
python-dotenv